requirements.txt
scipy
opencv-python
joblib
laspy
jaxtyping
packaging
torch_scatter
debugpy
wheel
triton
lightning
black
ruff
hydra-core
jaxtyping
beartype
colorama
scikit-image
colorspacious
matplotlib
moviepy==1.0.3
imageio
timm
DateTime
imgaug
iopath
imagecorruptions
mmcv
tensorboardX
HTML4Vision
dacite
lpips
e3nn
plyfile
tabulate
svg.py
scikit-video
Pillow
huggingface_hub
gradio
pydantic
open3d==0.19.0
einops
safetensors
tqdm
plyfile
roma
./pytorch3d